Move common inspection logic to one file
