+++ libvast/include/vast/fwd.hpp
@@ -15,6 +15,7 @@
 #include <caf/type_id.hpp>
 
 #include <cstdint>
+#include <filesystem>
 #include <map>
 #include <unordered_map>
 #include <vector>
@@ -80,6 +81,31 @@ class inbound_stream_slot;
 
 template <class, class...>
 class outbound_stream_slot;
+
+// TODO CAF 0.19. Check if this already implemented by CAF itself.
+template <class Slot>
+struct inspector_access<inbound_stream_slot<Slot>> {
+  template <class Inspector, class T>
+  static auto apply(Inspector& f, inbound_stream_slot<T>& x) {
+    auto val = x.value();
+    auto result = f.apply(val);
+    if constexpr (Inspector::is_loading)
+      x = inbound_stream_slot<T>{val};
+    return result;
+  }
+};
+
+template <>
+struct inspector_access<std::filesystem::path> {
+  template <class Inspector>
+  static auto apply(Inspector& f, std::filesystem::path& x) {
+    auto str = x.string();
+    auto result = f.apply(str);
+    if constexpr (Inspector::is_loading)
+      x = {str};
+    return result;
+  }
+};
 
 namespace detail {
